+++ src/probnum/diffeq/odefiltsmooth/__init__.py
@@ -1,9 +1,6 @@
 """ODE Filtering."""
 
-from .initialize import (
-    initialize_odefilter_with_rk,
-    initialize_odefilter_with_taylormode,
-)
-from .ivpfiltsmooth import GaussianIVPFilter
-from .kalman_odesolution import KalmanODESolution
-from .odefiltsmooth import probsolve_ivp
+from ._ivpfiltsmooth import GaussianIVPFilter
+from ._kalman_odesolution import KalmanODESolution
+
+__all__ = ["GaussianIVPFilter", "KalmanODESolution"]
