+++ src/probnum/diffeq/odefiltsmooth/initialize/__init__.py
@@ -1,0 +1,6 @@
+"""Initialisation procedures for ODE filters."""
+
+from ._initialize_with_runge_kutta import initialize_odefilter_with_rk
+from ._initialize_with_taylormode import initialize_odefilter_with_taylormode
+
+__all__ = ["initialize_odefilter_with_rk", "initialize_odefilter_with_taylormode"]
