+++ src/probnum/diffeq/perturbed/_perturbedstepsolver.py
@@ -6,9 +6,9 @@ import numpy as np
 
 from probnum import _randomvariablelist, randvars
 from probnum.typing import FloatArgType
+from src.probnum.diffeq.scipy.wrappedscipysolver import WrappedScipyRungeKutta
 
 from ..odesolver import ODESolver
-from ..wrappedscipysolver import WrappedScipyRungeKutta
 from ._perturbedstepsolution import PerturbedStepSolution
