+++ src/probnum/diffeq/stepsize/__init__.py
@@ -1,0 +1,12 @@
+"""Step-size selection rules."""
+
+from ._steprule import AdaptiveSteps, ConstantSteps, StepRule, propose_firststep
+
+__all__ = ["StepRule", "AdaptiveSteps", "ConstantSteps", "propose_firststep"]
+
+
+# Set correct module paths (for superclasses).
+# Corrects links and module paths in documentation.
+StepRule.__module__ = "probnum.diffeq.stepsize"
+AdaptiveSteps.__module__ = "probnum.diffeq.stepsize"
+ConstantSteps.__module__ = "probnum.diffeq.stepsize"
